1175-1182

+++ 1151_FibonacciFacil.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 
 int main(){
-    int prev_i = 1, tmp, c = 0, n;
-    scanf("%d", &n);
+    long int prev_i = 1, tmp, c = 0, n;
+    scanf("%ld", &n);
 
-    for(int i = 0; ; i += tmp){
+    for(long int i = 0; ; i += tmp){
         if(c == n){ break; }
-        printf("%d%s", i, c+1 == n ? "\n":" ");
+        printf("%ld%s", i, c+1 == n ? "\n":" ");
         tmp = prev_i;
         prev_i = i;
         c++;
